docs: update README
