+++ src/defaultContentTypes.ts
@@ -1,0 +1,18 @@
+export const defaultContentTypes = {
+    'image/png': ['.png'],
+    'image/jpeg': ['.jpg', '.jpeg'],
+    'image/svg+xml': ['.svg'],
+    'image/tiff': ['.tiff'],
+    'image/heif': ['.heif'],
+    'image/heic': ['.heic'],
+    'text/javascript': ['.js'],
+    'text/css': ['.css'],
+    'font/otf': ['.otf'],
+    'font/ttf': ['.ttf'],
+    'font/woff': ['.woff'],
+    'font/woff2': ['.woff2'],
+    'text/html': ['.html', '.htm'],
+    'text/plain': ['.txt', '.log'],
+    'application/json': ['.json'],
+    'image/vnd.microsoft.icon': ['.ico'],
+};
