+++ src/getTools.ts
@@ -10,8 +10,7 @@ import {IOptions} from './types';
 export const getTools = (
     options: IOptions = {},
 ) => {
-    const prefix = `/${options.scriptPrefix || 'middleware-static-livereload'}`.replace(/^\/+/, '/');
-    const clientScriptPath = `${prefix}.js`;
+    const clientScriptPath = `/${options.scriptPath || 'middleware-static-livereload.js'}`.replace(/^\/+/, '/');
     const findFile = createFileFinder(options, {
         [clientScriptPath]: path.join(__dirname, 'client-script.js'),
         [`${clientScriptPath}/polyfill.js`]: require.resolve('event-source-polyfill/src/eventsource.min.js'),
